sdn-demo/src/requirements.txt
cloudshell-core>=2.2,<2.3
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-networking-devices>=2.3,<2.4
